requirements.txt
# Django Backend
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
django-storages==1.14.2
boto3==1.28.85

# Telegram Bot
aiogram==3.2.0
aiohttp==3.8.6

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7

# Database
django-extensions==3.2.3

# Development
django-debug-toolbar==4.2.0
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

